requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi @ file:///C:/b/abs_4f5wo627a3/croots/recipe/certifi_1663615677642/work/certifi
chardet==5.0.0
charset-normalizer==2.1.1
click==8.1.3
cmake==3.24.1.1
colorama==0.4.5
decorator==4.4.2
Django==4.0
djangorestframework==3.14.0
dlib==19.24.0
face-recognition==1.3.0
face-recognition-models==0.3.0
idna==3.4
imageio==2.22.0
imageio-ffmpeg==0.4.7
moviepy==1.0.3
mysqlclient==2.1.1
numpy==1.23.3
opencv-python==4.6.0.66
Pillow==9.2.0
proglog==0.1.10
psycopg2==2.9.4
pytz==2022.4
requests==2.28.1
scipy==1.9.1
sqlparse==0.4.3
tqdm==4.64.1
tzdata==2022.5
urllib3==1.26.12
wincertstore==0.2
